+++ build.gradle
@@ -196,7 +196,7 @@ dependencies {
   implementation group: 'com.opencsv', name: 'opencsv', version: '5.7.1'
   implementation group: 'commons-validator', name: 'commons-validator', version: '1.7'
 
-  implementation group: 'com.github.hmcts', name: 'pip-data-models', version: '1.4.4', {
+  implementation group: 'com.github.hmcts', name: 'pip-data-models', version: '1.4.5', {
     exclude group: 'org.springframework.boot', module: 'spring-boot-starter-data-jpa'
   }
