+++ src/functionalTest/java/uk/gov/hmcts/reform/pip/account/management/controllers/AuditTest.java
@@ -20,8 +20,8 @@ import uk.gov.hmcts.reform.pip.account.management.config.AzureConfigurationClien
 import uk.gov.hmcts.reform.pip.account.management.model.AuditLog;
 import uk.gov.hmcts.reform.pip.account.management.model.AuditLogDto;
 import uk.gov.hmcts.reform.pip.account.management.model.CustomPageImpl;
-import uk.gov.hmcts.reform.pip.account.management.model.Roles;
-import uk.gov.hmcts.reform.pip.account.management.model.UserProvenances;
+import uk.gov.hmcts.reform.pip.model.account.Roles;
+import uk.gov.hmcts.reform.pip.model.account.UserProvenances;
 import uk.gov.hmcts.reform.pip.model.enums.AuditAction;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
