+++ src/main/java/uk/gov/hmcts/reform/pip/account/management/model/AuditLogDto.java
@@ -3,6 +3,8 @@ package uk.gov.hmcts.reform.pip.account.management.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import uk.gov.hmcts.reform.pip.account.management.validation.annotations.ValueOfEnum;
+import uk.gov.hmcts.reform.pip.model.account.Roles;
+import uk.gov.hmcts.reform.pip.model.account.UserProvenances;
 import uk.gov.hmcts.reform.pip.model.enums.AuditAction;
 
 import javax.validation.constraints.Email;
