--- src/main/java/uk/gov/hmcts/reform/pip/account/management/model/Sensitivity.java
@@ -1,8 +1,0 @@
-package uk.gov.hmcts.reform.pip.account.management.model;
-
-/**
- * A class which defines the sensitivity of the publication.
- */
-public enum Sensitivity {
-    PUBLIC, CLASSIFIED, INTERNAL, PRIVATE
-}
