+++ src/main/java/uk/gov/hmcts/reform/pip/account/management/service/SubscriptionService.java
@@ -3,14 +3,14 @@ package uk.gov.hmcts.reform.pip.account.management.service;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClientException;
 
 import static org.springframework.security.oauth2.client.web.reactive.function.client.ServerOAuth2AuthorizedClientExchangeFilterFunction.clientRegistrationId;
 
 @Slf4j
-@Component
+@Service
 /**
  * Service to communicate with subscription management microservice.
  */
