+++ src/main/java/uk/gov/hmcts/reform/pip/account/management/validation/validator/ConditionalEmailValidator.java
@@ -2,8 +2,8 @@ package uk.gov.hmcts.reform.pip.account.management.validation.validator;
 
 import org.apache.commons.validator.routines.EmailValidator;
 import uk.gov.hmcts.reform.pip.account.management.model.PiUser;
-import uk.gov.hmcts.reform.pip.account.management.model.Roles;
 import uk.gov.hmcts.reform.pip.account.management.validation.annotations.PiEmailConditionalValidation;
+import uk.gov.hmcts.reform.pip.model.account.Roles;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
