+++ src/test/java/uk/gov/hmcts/reform/pip/account/management/model/SystemAdminAccountTest.java
@@ -2,6 +2,8 @@ package uk.gov.hmcts.reform.pip.account.management.model;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import uk.gov.hmcts.reform.pip.model.account.Roles;
+import uk.gov.hmcts.reform.pip.model.account.UserProvenances;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
