+++ src/test/java/uk/gov/hmcts/reform/pip/account/management/service/AccountServiceTest.java
@@ -16,12 +16,12 @@ import uk.gov.hmcts.reform.pip.account.management.errorhandling.exceptions.UserN
 import uk.gov.hmcts.reform.pip.account.management.errorhandling.exceptions.UserWithProvenanceNotFoundException;
 import uk.gov.hmcts.reform.pip.account.management.model.AzureAccount;
 import uk.gov.hmcts.reform.pip.account.management.model.CreationEnum;
-import uk.gov.hmcts.reform.pip.account.management.model.ListType;
 import uk.gov.hmcts.reform.pip.account.management.model.PiUser;
-import uk.gov.hmcts.reform.pip.account.management.model.Roles;
-import uk.gov.hmcts.reform.pip.account.management.model.Sensitivity;
-import uk.gov.hmcts.reform.pip.account.management.model.UserProvenances;
 import uk.gov.hmcts.reform.pip.account.management.model.errored.ErroredPiUser;
+import uk.gov.hmcts.reform.pip.model.account.Roles;
+import uk.gov.hmcts.reform.pip.model.account.UserProvenances;
+import uk.gov.hmcts.reform.pip.model.publication.ListType;
+import uk.gov.hmcts.reform.pip.model.publication.Sensitivity;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -44,7 +44,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.reform.pip.account.management.model.Roles.SYSTEM_ADMIN;
+import static uk.gov.hmcts.reform.pip.model.account.Roles.SYSTEM_ADMIN;
 
 @ExtendWith(MockitoExtension.class)
 @SuppressWarnings("PMD.TooManyMethods")
@@ -133,7 +133,7 @@ class AccountServiceTest {
                                   Roles.INTERNAL_ADMIN_CTSC, FORENAME, SURNAME, null, null, null);
         PiUser user2 = new PiUser(UUID.randomUUID(), UserProvenances.PI_AAD, "567", "test@test.com",
 
-                                 Roles.INTERNAL_ADMIN_CTSC, FORENAME, SURNAME, null, null, null);
+                                  Roles.INTERNAL_ADMIN_CTSC, FORENAME, SURNAME, null, null, null);
         List<PiUser> users = new ArrayList<>();
         users.add(user1);
         users.add(user2);
